separate hardware related and non-releated code unit test binaries no go into  test/bin/ folder
