+++ src/jled_base.cpp
@@ -19,8 +19,7 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
 // IN THE SOFTWARE.
 //
-#include <Arduino.h>
-#include <jled.h>
+#include "jled_base.h"  // NOLINT
 
 
 namespace jled {
